lakeishas_cupcakery/requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.3
django-environ>=0.11.2

# Authentication
djangorestframework-simplejwt>=5.3.0
python-jose>=3.3.0
python-jose[cryptography]>=3.3.0

# Database
psycopg2-binary>=2.9.9

# Email
sendgrid>=6.11.0
python-dotenv>=1.0.0

# Media Storage
Pillow>=10.0.0
whitenoise>=6.5.0

# API Documentation
drf-yasg>=1.21.5

# Testing
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
djlint>=1.32.0

# Development
django-debug-toolbar>=4.1.0
python-dotenv>=1.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0

# Security
python-decouple>=3.8
django-ratelimit>=3.0.1
